before change to User Repo

+++ app/src/main/java/com/example/khajakhoj/model/Review.kt
@@ -1,10 +1,10 @@
 package com.example.khajakhoj.model
 
 data class Review(
-    val id: String = "",
-    val userId: String = "",
-    val restaurantId: String = "",
-    val rating: Double = 0.0,
+    val id: String = "", // reviewId
+    val userId: String = "", // id of user s who wrote the review
+    val restaurantId: String = "", // id of restaurant
+    val rating: Int = 0,
     val comment: String = "",
     val createdAt: Long = 0L
 )
